[rdrawable] extract RPadBase into separate source files
