+++ graf2d/gpadv7/CMakeLists.txt
@@ -29,6 +29,7 @@ ROOT_STANDARD_LIBRARY_PACKAGE(ROOTGpadv7
     ROOT/RStyle.hxx
     ROOT/RPadDisplayItem.hxx
     ROOT/RPadExtent.hxx
+    ROOT/RPadBase.hxx
     ROOT/RPad.hxx
     ROOT/RPadLength.hxx
     ROOT/RPadPainter.hxx
@@ -48,6 +49,7 @@ ROOT_STANDARD_LIBRARY_PACKAGE(ROOTGpadv7
     v7/src/RPalette.cxx
     v7/src/RStyle.cxx
     v7/src/RStyleReader.cxx
+    v7/src/RPadBase.cxx
     v7/src/RPad.cxx
     v7/src/RPadExtent.cxx
     v7/src/RPadLength.cxx
@@ -61,6 +63,5 @@ ROOT_STANDARD_LIBRARY_PACKAGE(ROOTGpadv7
     Core
     RIO
 )
-
 
 ROOT_ADD_TEST_SUBDIRECTORY(v7/test)
